Extract file processing code
